Add env_config and config_prefix options for config.properties alternative
Setting many kafka parameters is not possible using environment variables in Linux because Java is unable to read environment varaibles
with '.' in them. Environment variables are supposed to be alphanumeric with underscores allowed only. Settings like kafka.bootstrap.servers
can only be set in the config.properties. Some organizations prefer configuration via environment variables to prevent seeing sensitive
data in `ps` or the docker image itself.

This change adds in a JSON-encoded environment variable that can be used as an alternative to config.properties. The input is expected to be
a JSON object of string keys and string values. With this change, we are able to set all necessarty configs in environment variables.

Additionally, this adds a `config_prefix` option that can additionally filter keys out of the config.properties or env_config.